requirements.txt
Flask==1.1.2
Flask-Cors==3.0.8
onnx~=1.5.0
onnxruntime~=1.2
pyyaml~=5.1.0
